Envio de correos al modificar libros

+++ app/Http/Controllers/BookController.php
@@ -77,6 +77,8 @@ class BookController extends Controller
         $book->category_id = $request->post('category');
         $book->book_image = $name;
 
+        $book->sendEmail($book, 'Se creo este libro:'); // Enviar email
+
         $book->save();
 
         return redirect('/');
@@ -146,6 +148,8 @@ class BookController extends Controller
         $book->book_description = $request->post('description');
         $book->category_id = $request->post('category');
 
+        $book->sendEmail($book, 'Se edito este libro:'); // Enviar email
+
         $book->save();
 
         return redirect('/category/show/tables/'.$book->category_id);
@@ -162,10 +166,21 @@ class BookController extends Controller
         $id = $request->post('book_id');
         $book = Book::find($id);
 
-        $file_path = public_path().'/images/'.$book->book_image;
-        \File::delete($file_path);
+        if($book->book_image != "pordefecto") {
+            $file_path = public_path().'/images/'.$book->book_image;
+            \File::delete($file_path);
+        }
+
+        $book->sendEmail($book, 'Se elimino este libro:'); // Enviar email
 
         $book->delete();
+        return back();
+    }
+
+    public function sendEmail(Book $book) {
+        Mail::to('luis.380171120858@ucaldas.edu.co')
+        ->send(new BookReport($book, $titulo));
+
         return back();
     }
 }
